faas/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-core==1.32.0
azure-functions==1.21.3
azure-storage-blob==12.24.1
bcrypt==3.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==41.0.7
idna==3.10
isodate==0.7.2
pillow==11.1.0
pycparser==2.22
requests==2.32.3
six==1.17.0
typing-extensions==4.12.2
urllib3==2.3.0